pull out the remaining tests
