+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/optimizer/rules/LiteralsOnTheRightTests.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+package org.elasticsearch.xpack.esql.optimizer.rules;
+
+import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.xpack.esql.core.expression.Alias;
+import org.elasticsearch.xpack.esql.core.expression.Expression;
+import org.elasticsearch.xpack.esql.core.expression.Literal;
+import org.elasticsearch.xpack.esql.expression.predicate.operator.comparison.Equals;
+
+import static org.elasticsearch.xpack.esql.EsqlTestUtils.FIVE;
+import static org.elasticsearch.xpack.esql.EsqlTestUtils.equalsOf;
+import static org.elasticsearch.xpack.esql.core.tree.Source.EMPTY;
+import static org.elasticsearch.xpack.esql.core.type.DataType.INTEGER;
+
+public class LiteralsOnTheRightTests extends ESTestCase {
+
+    public void testLiteralsOnTheRight() {
+        Alias a = new Alias(EMPTY, "a", new Literal(EMPTY, 10, INTEGER));
+        Expression result = new LiteralsOnTheRight().rule(equalsOf(FIVE, a));
+        assertTrue(result instanceof Equals);
+        Equals eq = (Equals) result;
+        assertEquals(a, eq.left());
+        assertEquals(FIVE, eq.right());
+
+        // Note: Null Equals test removed here
+    }
+
+}
